+++ src/main/java/http/common/HttpHeader.java
@@ -1,4 +1,4 @@
-package network;
+package http.common;
 
 import java.util.Collections;
 import java.util.List;
