+++ src/main/java/http/exception/IllegalHttpVersionException.java
@@ -1,4 +1,4 @@
-package network.exception;
+package http.exception;
 
 public class IllegalHttpVersionException extends RuntimeException {
     public static final String MESSAGE = "올바르지 않은 Http 버전입니다.";
