+++ src/main/java/http/request/HttpMethod.java
@@ -1,4 +1,4 @@
-package network;
+package http.request;
 
 public enum HttpMethod {
     GET, POST, PUT, DELETE
