+++ src/main/java/http/request/HttpRequestParams.java
@@ -1,4 +1,4 @@
-package network;
+package http.request;
 
 import java.util.Arrays;
 import java.util.Collections;
