+++ src/main/java/http/response/HttpResponseConverter.java
@@ -1,4 +1,4 @@
-package network;
+package http.response;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
