+++ src/main/java/webserver/WebServer.java
@@ -1,10 +1,10 @@
 package webserver;
 
-import java.net.ServerSocket;
-import java.net.Socket;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.net.ServerSocket;
+import java.net.Socket;
 
 public class WebServer {
     private static final Logger logger = LoggerFactory.getLogger(WebServer.class);
