--- src/test/java/network/HttpRequestTest.java
@@ -1,7 +1,0 @@
-package network;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class HttpRequestTest {
-
-}
